TestPerformed view corrected

+++ database/migrations/2021_05_11_120523_create_available_test_inventories_table.php
@@ -17,7 +17,7 @@ class CreateAvailableTestInventoriesTable extends Migration
             $table->bigIncrements('id');
             $table->integer('available_test_id');
             $table->string('inventory_id');
-            $table->integer('itemUsed');
+            $table->bigInteger('itemUsed');
             $table->timestamps();
         });
     }
